QProcess execute Answer program and make output text files

+++ CodeScoring/mainwindow.cpp
@@ -112,27 +112,89 @@ void MainWindow::on_answerCombo_activated(int index)
 
 void MainWindow::on_ScoringButton_clicked()
 {
+    //check init
     if(!isInitialized){
         QString ini="Open Folder and Load it first";
         QMessageBox::warning(0,"No Initializing Error",ini);
         return;
     }
 
-    if(!gnu_process){
-        gnu_process= new QProcess(this);
-    }
-    QString program = "g++";
-    QStringList command;
-    command << "-std=c++14"<<outputFileList.at(0).filePath()<<"-o"<<temp_execute_folder.absolutePath()+"/"+outputFileList.at(0).baseName();
+    ui->ScoringButton->setEnabled(false);
+
     QProcessEnvironment env= QProcessEnvironment::systemEnvironment();
-    gnu_process->setProcessEnvironment(env);
-    gnu_process->setProcessChannelMode(QProcess::MergedChannels);
-    gnu_process->start(program,command);
-    gnu_process->waitForFinished();
-    QString output(gnu_process->readAllStandardOutput());
-    //QMessageBox::warning(0,"output", output);
-    gnu_process->terminate();
-    //int i=myprocess->exitCode();
+
+
+    int answerCompileOption = ui ->answerEnvironment->currentIndex();
+    QString program;
+    QString stdoption;
+    QString outFileName;
+    QStringList command;
+
+    //make Answer output
+    if(ui->answerCombo->currentIndex()==0){
+        command.clear();
+        outFileName=answer_output_folder.absolutePath()+"/"+outputFileList.at(0).baseName();
+
+        switch(answerCompileOption){
+        case 0: program="gcc"; command << outputFileList.at(0).filePath()<<"-o"<<outFileName; break;
+        case 1: program="gcc"; stdoption="-std=c99"; command << stdoption << outputFileList.at(0).filePath()<<"-o"<<outFileName; break;
+        case 2: program="g++"; command << outputFileList.at(0).filePath()<<"-o"<<outFileName; break;
+        case 3: program="g++"; stdoption="-std=c++11"; command << stdoption << outputFileList.at(0).filePath()<<"-o"<<outFileName; break;
+        case 4: program="g++"; stdoption="-std=c++14"; command << stdoption << outputFileList.at(0).filePath()<<"-o"<<outFileName; break;
+        }
+
+        if(!gnu_process){
+            gnu_process=new QProcess(this);
+        }
+
+        gnu_process->setProcessEnvironment(env);
+        gnu_process->setProcessChannelMode(QProcess::MergedChannels);
+        gnu_process->start(program,command);
+        ui->ProcessingMessage->setText("Compiling Answer Code...");
+        gnu_process->waitForFinished();
+        ui->ProcessingMessage->setText("Compiling Answer Code Compeleted");
+        gnu_process->terminate();
+
+        program="./"+outFileName;
+        for(int i=0;i<inputFileList.size();i++){
+            ui->ProcessingMessage->setText("Processing "+inputFileList.at(i).baseName());
+
+            command.clear();
+            QString savedfileName=inputFileList.at(i).baseName()+"out.txt";
+            //command << "<"+inputFileList.at(i).filePath() << ">"+answer_output_folder.absolutePath()+"/"+savedfileName;
+
+            QMessageBox::warning(0,"output", "<"+inputFileList.at(i).filePath());
+
+            gnu_process->setProcessChannelMode(QProcess::SeparateChannels);
+            gnu_process->setStandardInputFile(inputFileList.at(i).filePath());
+            gnu_process->setStandardOutputFile(answer_output_folder.absolutePath()+"/"+savedfileName);
+            //gnu_process->set;
+            gnu_process->start(program);
+            gnu_process->waitForFinished();
+            gnu_process->terminate();
+        }
+        ui->ProcessingMessage->setText("Complete to make Answer Files");
+
+    }
+
+
+//    if(!gnu_process){
+//        gnu_process= new QProcess(this);
+//    }
+//    QString program = "g++";
+//    QStringList command;
+//    command << "-std=c++14"<<outputFileList.at(0).filePath()<<"-o"<<temp_execute_folder.absolutePath()+"/"+outputFileList.at(0).baseName();
+//    QProcessEnvironment env= QProcessEnvironment::systemEnvironment();
+//    gnu_process->setProcessEnvironment(env);
+//    gnu_process->setProcessChannelMode(QProcess::MergedChannels);
+//    gnu_process->start(program,command);
+//    gnu_process->waitForFinished();
+//    QString output(gnu_process->readAllStandardOutput());
+//    //QMessageBox::warning(0,"output", output);
+//    gnu_process->terminate();
+//    //int i=myprocess->exitCode();
+
+    ui->ScoringButton->setEnabled(true);
 }
 
 void MainWindow::on_exitButton_clicked()
